+++ Source/Device.generated.swift
@@ -15,7 +15,7 @@ import WatchKit
 import UIKit
 #endif
 
-// MARK: - Device
+// MARK: Device
 
 /// This enum is a value-type wrapper and extension of
 /// [`UIDevice`](https://developer.apple.com/library/ios/documentation/UIKit/Reference/UIDevice_Class/).
@@ -144,6 +144,10 @@ public enum Device {
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP806/sp806-iphone11pro-max_2x.png)
     case iPhone11ProMax
+    /// Device is an [iPhone SE (2nd generation)](https://support.apple.com/kb/SP820)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP820/iphone-se-2nd-gen_2x.png)
+    case iPhoneSE2
     /// Device is an [iPad 2](https://support.apple.com/kb/SP622)
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP622/SP622_01-ipad2-mul.png)
@@ -180,6 +184,10 @@ public enum Device {
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP807/sp807-ipad-7th-gen_2x.png)
     case iPad7
+    /// Device is an [iPad (8th generation)](https://support.apple.com/kb/SP822)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP822/sp822-ipad-8gen_2x.png)
+    case iPad8
     /// Device is an [iPad Mini](https://support.apple.com/kb/SP661)
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP661/sp661_ipad_mini_color.jpg)
@@ -224,6 +232,14 @@ public enum Device {
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP785/ipad-pro-12-2018_2x.png)
     case iPadPro12Inch3
+    /// Device is an [iPad Pro 11-inch (2nd generation)](https://support.apple.com/kb/SP814)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP814/ipad-pro-11-2020.jpeg)
+    case iPadPro11Inch2
+    /// Device is an [iPad Pro 12.9-inch (4th generation)](https://support.apple.com/kb/SP815)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP815/ipad-pro-12-2020.jpeg)
+    case iPadPro12Inch4
     /// Device is a [HomePod](https://support.apple.com/kb/SP773)
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP773/homepod_space_gray_large_2x.jpg)
@@ -232,7 +248,7 @@ public enum Device {
     /// Device is an [Apple TV HD](https://support.apple.com/kb/SP724) (Previously Apple TV (4th generation))
     ///
     /// ![Image](http://images.apple.com/v/tv/c/images/overview/buy_tv_large_2x.jpg)
-    case appleTV4
+    case appleTVHD
     /// Device is an [Apple TV 4K](https://support.apple.com/kb/SP769)
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP769/appletv4k.png)
@@ -286,6 +302,22 @@ public enum Device {
     ///
     /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP808/sp808-apple-watch-series-5_2x.png)
     case appleWatchSeries5_44mm
+    /// Device is an [Apple Watch Series 6](https://support.apple.com/kb/SP826)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP826/sp826-apple-watch-series6-580_2x.png)
+    case appleWatchSeries6_40mm
+    /// Device is an [Apple Watch Series 6](https://support.apple.com/kb/SP826)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP826/sp826-apple-watch-series6-580_2x.png)
+    case appleWatchSeries6_44mm
+    /// Device is an [Apple Watch SE](https://support.apple.com/kb/SP827)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP827/sp827-apple-watch-se-580_2x.png)
+    case appleWatchSE_40mm
+    /// Device is an [Apple Watch SE](https://support.apple.com/kb/SP827)
+    ///
+    /// ![Image](https://support.apple.com/library/APPLE/APPLECARE_ALLGEOS/SP827/sp827-apple-watch-se-580_2x.png)
+    case appleWatchSE_44mm
   #endif
 
   /// Device is [Simulator](https://developer.apple.com/library/ios/documentation/IDEs/Conceptual/iOS_Simulator_Guide/Introduction/Introduction.html)
@@ -348,6 +380,7 @@ public enum Device {
       case "iPhone12,1": return iPhone11
       case "iPhone12,3": return iPhone11Pro
       case "iPhone12,5": return iPhone11ProMax
+      case "iPhone12,8": return iPhoneSE2
       case "iPad2,1", "iPad2,2", "iPad2,3", "iPad2,4": return iPad2
       case "iPad3,1", "iPad3,2", "iPad3,3": return iPad3
       case "iPad3,4", "iPad3,5", "iPad3,6": return iPad4
@@ -357,6 +390,7 @@ public enum Device {
       case "iPad7,5", "iPad7,6": return iPad6
       case "iPad11,3", "iPad11,4": return iPadAir3
       case "iPad7,11", "iPad7,12": return iPad7
+      case "iPad11,6", "iPad11,7": return iPad8
       case "iPad2,5", "iPad2,6", "iPad2,7": return iPadMini
       case "iPad4,4", "iPad4,5", "iPad4,6": return iPadMini2
       case "iPad4,7", "iPad4,8", "iPad4,9": return iPadMini3
@@ -368,13 +402,15 @@ public enum Device {
       case "iPad7,3", "iPad7,4": return iPadPro10Inch
       case "iPad8,1", "iPad8,2", "iPad8,3", "iPad8,4": return iPadPro11Inch
       case "iPad8,5", "iPad8,6", "iPad8,7", "iPad8,8": return iPadPro12Inch3
+      case "iPad8,9", "iPad8,10": return iPadPro11Inch2
+      case "iPad8,11", "iPad8,12": return iPadPro12Inch4
       case "AudioAccessory1,1": return homePod
       case "i386", "x86_64": return simulator(mapToDevice(identifier: ProcessInfo().environment["SIMULATOR_MODEL_IDENTIFIER"] ?? "iOS"))
       default: return unknown(identifier)
       }
     #elseif os(tvOS)
       switch identifier {
-      case "AppleTV5,3": return appleTV4
+      case "AppleTV5,3": return appleTVHD
       case "AppleTV6,2": return appleTV4K
       case "i386", "x86_64": return simulator(mapToDevice(identifier: ProcessInfo().environment["SIMULATOR_MODEL_IDENTIFIER"] ?? "tvOS"))
       default: return unknown(identifier)
@@ -393,6 +429,10 @@ public enum Device {
       case "Watch4,2", "Watch4,4": return appleWatchSeries4_44mm
       case "Watch5,1", "Watch5,3": return appleWatchSeries5_40mm
       case "Watch5,2", "Watch5,4": return appleWatchSeries5_44mm
+      case "Watch6,1", "Watch6,3": return appleWatchSeries6_40mm
+      case "Watch6,2", "Watch6,4": return appleWatchSeries6_44mm
+      case "Watch5,9", "Watch5,11": return appleWatchSE_40mm
+      case "Watch5,10", "Watch5,12": return appleWatchSE_44mm
       case "i386", "x86_64": return simulator(mapToDevice(identifier: ProcessInfo().environment["SIMULATOR_MODEL_IDENTIFIER"] ?? "watchOS"))
       default: return unknown(identifier)
       }
@@ -443,6 +483,7 @@ public enum Device {
         case .iPhone11: return 6.1
         case .iPhone11Pro: return 5.8
         case .iPhone11ProMax: return 6.5
+        case .iPhoneSE2: return 4.7
         case .iPad2: return 9.7
         case .iPad3: return 9.7
         case .iPad4: return 9.7
@@ -452,6 +493,7 @@ public enum Device {
         case .iPad6: return 9.7
         case .iPadAir3: return 10.5
         case .iPad7: return 10.2
+        case .iPad8: return 10.2
         case .iPadMini: return 7.9
         case .iPadMini2: return 7.9
         case .iPadMini3: return 7.9
@@ -463,6 +505,8 @@ public enum Device {
         case .iPadPro10Inch: return 10.5
         case .iPadPro11Inch: return 11.0
         case .iPadPro12Inch3: return 12.9
+        case .iPadPro11Inch2: return 11.0
+        case .iPadPro12Inch4: return 12.9
         case .homePod: return -1
         case .simulator(let model): return model.diagonal
         case .unknown: return -1
@@ -481,6 +525,10 @@ public enum Device {
       case .appleWatchSeries4_44mm: return 2.0
       case .appleWatchSeries5_40mm: return 1.8
       case .appleWatchSeries5_44mm: return 2.0
+      case .appleWatchSeries6_40mm: return 1.8
+      case .appleWatchSeries6_44mm: return 2.0
+      case .appleWatchSE_40mm: return 1.8
+      case .appleWatchSE_44mm: return 2.0
       case .simulator(let model): return model.diagonal
       case .unknown: return -1
       }
@@ -516,6 +564,7 @@ public enum Device {
       case .iPhone11: return (width: 9, height: 19.5)
       case .iPhone11Pro: return (width: 9, height: 19.5)
       case .iPhone11ProMax: return (width: 9, height: 19.5)
+      case .iPhoneSE2: return (width: 9, height: 16)
       case .iPad2: return (width: 3, height: 4)
       case .iPad3: return (width: 3, height: 4)
       case .iPad4: return (width: 3, height: 4)
@@ -525,6 +574,7 @@ public enum Device {
       case .iPad6: return (width: 3, height: 4)
       case .iPadAir3: return (width: 3, height: 4)
       case .iPad7: return (width: 3, height: 4)
+      case .iPad8: return (width: 3, height: 4)
       case .iPadMini: return (width: 3, height: 4)
       case .iPadMini2: return (width: 3, height: 4)
       case .iPadMini3: return (width: 3, height: 4)
@@ -536,6 +586,8 @@ public enum Device {
       case .iPadPro10Inch: return (width: 3, height: 4)
       case .iPadPro11Inch: return (width: 139, height: 199)
       case .iPadPro12Inch3: return (width: 512, height: 683)
+      case .iPadPro11Inch2: return (width: 139, height: 199)
+      case .iPadPro12Inch4: return (width: 512, height: 683)
       case .homePod: return (width: 4, height: 5)
       case .simulator(let model): return model.screenRatio
       case .unknown: return (width: -1, height: -1)
@@ -554,6 +606,10 @@ public enum Device {
       case .appleWatchSeries4_44mm: return (width: 4, height: 5)
       case .appleWatchSeries5_40mm: return (width: 4, height: 5)
       case .appleWatchSeries5_44mm: return (width: 4, height: 5)
+      case .appleWatchSeries6_40mm: return (width: 4, height: 5)
+      case .appleWatchSeries6_44mm: return (width: 4, height: 5)
+      case .appleWatchSE_40mm: return (width: 4, height: 5)
+      case .appleWatchSE_44mm: return (width: 4, height: 5)
       case .simulator(let model): return model.screenRatio
       case .unknown: return (width: -1, height: -1)
       }
@@ -570,27 +626,28 @@ public enum Device {
 
     /// All iPhones
     public static var allPhones: [Device] {
-      return [.iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax]
+      return [.iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2]
     }
 
     /// All iPads
     public static var allPads: [Device] {
-      return [.iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
+      return [.iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4]
     }
 
     /// All X-Series Devices
+    @available(*, deprecated, renamed: "allDevicesWithSensorHousing")
     public static var allXSeriesDevices: [Device] {
       return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax]
     }
 
-    /// All Plus-Sized Devices
+    /// All Plus and Max-Sized Devices
     public static var allPlusSizedDevices: [Device] {
       return [.iPhone6Plus, .iPhone6sPlus, .iPhone7Plus, .iPhone8Plus, .iPhoneXSMax, .iPhone11ProMax]
     }
 
     /// All Pro Devices
     public static var allProDevices: [Device] {
-      return [.iPhone11Pro, .iPhone11ProMax, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
+      return [.iPhone11Pro, .iPhone11ProMax, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4]
     }
 
     /// All mini Devices
@@ -618,12 +675,13 @@ public enum Device {
       return allMiniDevices.map(Device.simulator)
     }
 
-    /// All simulator Plus-Sized Devices
+    /// All simulator X series Devices
+    @available(*, deprecated, renamed: "allSimulatorDevicesWithSensorHousing")
     public static var allSimulatorXSeriesDevices: [Device] {
-      return allXSeriesDevices.map(Device.simulator)
+      return allDevicesWithSensorHousing.map(Device.simulator)
     }
 
-    /// All simulator Plus-Sized Devices
+    /// All simulator Plus and Max-Sized Devices
     public static var allSimulatorPlusSizedDevices: [Device] {
       return allPlusSizedDevices.map(Device.simulator)
     }
@@ -676,22 +734,27 @@ public enum Device {
 
     /// All Touch ID Capable Devices
     public static var allTouchIDCapableDevices: [Device] {
-      return [.iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch]
+      return [.iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneSE2, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch]
     }
 
     /// All Face ID Capable Devices
     public static var allFaceIDCapableDevices: [Device] {
-      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPadPro11Inch, .iPadPro12Inch3]
+      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4]
+    }
+
+    /// All Devices with Touch ID or Face ID
+    public static var allBiometricAuthenticationCapableDevices: [Device] {
+      return [.iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4]
     }
 
     /// Returns whether or not the device has Touch ID
     public var isTouchIDCapable: Bool {
-      return isOneOf(Device.allTouchIDCapableDevices)
+      return isOneOf(Device.allTouchIDCapableDevices) || isOneOf(Device.allTouchIDCapableDevices.map(Device.simulator))
     }
 
     /// Returns whether or not the device has Face ID
     public var isFaceIDCapable: Bool {
-      return isOneOf(Device.allFaceIDCapableDevices)
+      return isOneOf(Device.allFaceIDCapableDevices) || isOneOf(Device.allFaceIDCapableDevices.map(Device.simulator))
     }
 
     /// Returns whether or not the device has any biometric sensor (i.e. Touch ID or Face ID)
@@ -704,44 +767,59 @@ public enum Device {
       return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax]
     }
 
+    /// All simulator devices that feature a sensor housing in the screen
+    public static var allSimulatorDevicesWithSensorHousing: [Device] {
+      return allDevicesWithSensorHousing.map(Device.simulator)
+    }
+
     /// Returns whether or not the device has a sensor housing
     public var hasSensorHousing: Bool {
-      return isOneOf(Device.allDevicesWithSensorHousing)
+      return isOneOf(Device.allDevicesWithSensorHousing) || isOneOf(Device.allDevicesWithSensorHousing.map(Device.simulator))
     }
 
     /// All devices that feature a screen with rounded corners.
     public static var allDevicesWithRoundedDisplayCorners: [Device] {
-      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPadPro11Inch, .iPadPro12Inch3]
+      return [.iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4]
     }
 
     /// Returns whether or not the device has a screen with rounded corners.
     public var hasRoundedDisplayCorners: Bool {
-      return isOneOf(Device.allDevicesWithRoundedDisplayCorners)
+      return isOneOf(Device.allDevicesWithRoundedDisplayCorners) || isOneOf(Device.allDevicesWithRoundedDisplayCorners.map(Device.simulator))
     }
 
     /// All devices that have 3D Touch support.
     public static var allDevicesWith3dTouchSupport: [Device] {
-      return [.iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax]
+      return [.iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax]
     }
 
     /// Returns whether or not the device has 3D Touch support.
     public var has3dTouchSupport: Bool {
-      return isOneOf(Device.allDevicesWith3dTouchSupport)
+      return isOneOf(Device.allDevicesWith3dTouchSupport) || isOneOf(Device.allDevicesWith3dTouchSupport.map(Device.simulator))
     }
 
     /// All devices that support wireless charging.
     public static var allDevicesWithWirelessChargingSupport: [Device] {
-      return [.iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax]
+      return [.iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2]
     }
 
-    /// Returns whether or not the device supports wireless charging
+    /// Returns whether or not the device supports wireless charging.
     public var supportsWirelessCharging: Bool {
-      return isOneOf(Device.allDevicesWithWirelessChargingSupport)
+      return isOneOf(Device.allDevicesWithWirelessChargingSupport) || isOneOf(Device.allDevicesWithWirelessChargingSupport.map(Device.simulator))
+    }
+
+    /// All devices that have a LiDAR sensor.
+    public static var allDevicesWithALidarSensor: [Device] {
+      return [.iPadPro11Inch2, .iPadPro12Inch4]
+    }
+
+    /// Returns whether or not the device has a LiDAR sensor.
+    public var hasLidarSensor: Bool {
+      return isOneOf(Device.allDevicesWithALidarSensor) || isOneOf(Device.allDevicesWithALidarSensor.map(Device.simulator))
     }
   #elseif os(tvOS)
     /// All TVs
     public static var allTVs: [Device] {
-       return [.appleTV4, .appleTV4K]
+       return [.appleTVHD, .appleTV4K]
     }
 
     /// All simulator TVs
@@ -751,7 +829,7 @@ public enum Device {
   #elseif os(watchOS)
     /// All Watches
     public static var allWatches: [Device] {
-       return [.appleWatchSeries0_38mm, .appleWatchSeries0_42mm, .appleWatchSeries1_38mm, .appleWatchSeries1_42mm, .appleWatchSeries2_38mm, .appleWatchSeries2_42mm, .appleWatchSeries3_38mm, .appleWatchSeries3_42mm, .appleWatchSeries4_40mm, .appleWatchSeries4_44mm, .appleWatchSeries5_40mm, .appleWatchSeries5_44mm]
+       return [.appleWatchSeries0_38mm, .appleWatchSeries0_42mm, .appleWatchSeries1_38mm, .appleWatchSeries1_42mm, .appleWatchSeries2_38mm, .appleWatchSeries2_42mm, .appleWatchSeries3_38mm, .appleWatchSeries3_42mm, .appleWatchSeries4_40mm, .appleWatchSeries4_44mm, .appleWatchSeries5_40mm, .appleWatchSeries5_44mm, .appleWatchSeries6_40mm, .appleWatchSeries6_44mm, .appleWatchSE_40mm, .appleWatchSE_44mm]
     }
 
     /// All simulator Watches
@@ -766,7 +844,7 @@ public enum Device {
 
     /// Returns whether or not the device has Force Touch support.
     public var hasForceTouchSupport: Bool {
-      return isOneOf(Device.allWatchesWithForceTouchSupport)
+      return isOneOf(Device.allWatchesWithForceTouchSupport) || isOneOf(Device.allWatchesWithForceTouchSupport.map(Device.simulator))
     }
   #endif
 
@@ -818,6 +896,8 @@ public enum Device {
   public func isOneOf(_ devices: [Device]) -> Bool {
     return devices.contains(self)
   }
+
+  // MARK: Current Device
 
   /// Whether or not the current device is the current device.
   private var isCurrent: Bool {
@@ -902,6 +982,7 @@ public enum Device {
       case .iPhone11: return 326
       case .iPhone11Pro: return 458
       case .iPhone11ProMax: return 458
+      case .iPhoneSE2: return 326
       case .iPad2: return 132
       case .iPad3: return 264
       case .iPad4: return 264
@@ -911,6 +992,7 @@ public enum Device {
       case .iPad6: return 264
       case .iPadAir3: return 264
       case .iPad7: return 264
+      case .iPad8: return 264
       case .iPadMini: return 163
       case .iPadMini2: return 326
       case .iPadMini3: return 326
@@ -922,6 +1004,8 @@ public enum Device {
       case .iPadPro10Inch: return 264
       case .iPadPro11Inch: return 264
       case .iPadPro12Inch3: return 264
+      case .iPadPro11Inch2: return 264
+      case .iPadPro12Inch4: return 264
       case .homePod: return -1
       case .simulator(let model): return model.ppi
       case .unknown: return nil
@@ -940,6 +1024,10 @@ public enum Device {
     case .appleWatchSeries4_44mm: return 326
     case .appleWatchSeries5_40mm: return 326
     case .appleWatchSeries5_44mm: return 326
+    case .appleWatchSeries6_40mm: return 326
+    case .appleWatchSeries6_44mm: return 326
+    case .appleWatchSE_40mm: return 326
+    case .appleWatchSE_44mm: return 326
     case .simulator(let model): return model.ppi
     case .unknown: return nil
     }
@@ -971,7 +1059,7 @@ public enum Device {
   }
 }
 
-// MARK: - CustomStringConvertible
+// MARK: CustomStringConvertible
 extension Device: CustomStringConvertible {
 
   /// A textual representation of the device.
@@ -1002,6 +1090,7 @@ extension Device: CustomStringConvertible {
       case .iPhone11: return "iPhone 11"
       case .iPhone11Pro: return "iPhone 11 Pro"
       case .iPhone11ProMax: return "iPhone 11 Pro Max"
+      case .iPhoneSE2: return "iPhone SE (2nd generation)"
       case .iPad2: return "iPad 2"
       case .iPad3: return "iPad (3rd generation)"
       case .iPad4: return "iPad (4th generation)"
@@ -1011,6 +1100,7 @@ extension Device: CustomStringConvertible {
       case .iPad6: return "iPad (6th generation)"
       case .iPadAir3: return "iPad Air (3rd generation)"
       case .iPad7: return "iPad (7th generation)"
+      case .iPad8: return "iPad (8th generation)"
       case .iPadMini: return "iPad Mini"
       case .iPadMini2: return "iPad Mini 2"
       case .iPadMini3: return "iPad Mini 3"
@@ -1022,8 +1112,10 @@ extension Device: CustomStringConvertible {
       case .iPadPro10Inch: return "iPad Pro (10.5-inch)"
       case .iPadPro11Inch: return "iPad Pro (11-inch)"
       case .iPadPro12Inch3: return "iPad Pro (12.9-inch) (3rd generation)"
+      case .iPadPro11Inch2: return "iPad Pro (11-inch) (2nd generation)"
+      case .iPadPro12Inch4: return "iPad Pro (12.9-inch) (4th generation)"
       case .homePod: return "HomePod"
-      case .simulator(let model): return "Simulator (\(model))"
+      case .simulator(let model): return "Simulator (\(model.description))"
       case .unknown(let identifier): return identifier
       }
     #elseif os(watchOS)
@@ -1040,21 +1132,116 @@ extension Device: CustomStringConvertible {
       case .appleWatchSeries4_44mm: return "Apple Watch Series 4 44mm"
       case .appleWatchSeries5_40mm: return "Apple Watch Series 5 40mm"
       case .appleWatchSeries5_44mm: return "Apple Watch Series 5 44mm"
-      case .simulator(let model): return "Simulator (\(model))"
+      case .appleWatchSeries6_40mm: return "Apple Watch Series 6 40mm"
+      case .appleWatchSeries6_44mm: return "Apple Watch Series 6 44mm"
+      case .appleWatchSE_40mm: return "Apple Watch SE 40mm"
+      case .appleWatchSE_44mm: return "Apple Watch SE 44mm"
+      case .simulator(let model): return "Simulator (\(model.description))"
       case .unknown(let identifier): return identifier
       }
     #elseif os(tvOS)
       switch self {
-      case .appleTV4: return "Apple TV HD"
+      case .appleTVHD: return "Apple TV HD"
       case .appleTV4K: return "Apple TV 4K"
-      case .simulator(let model): return "Simulator (\(model))"
+      case .simulator(let model): return "Simulator (\(model.description))"
       case .unknown(let identifier): return identifier
       }
     #endif
   }
+
+  /// A safe version of `description`.
+  /// Example:
+  /// Device.iPhoneXR.description:     iPhone Xʀ
+  /// Device.iPhoneXR.safeDescription: iPhone XR
+  public var safeDescription: String {
+    #if os(iOS)
+      switch self {
+      case .iPodTouch5: return "iPod touch (5th generation)"
+      case .iPodTouch6: return "iPod touch (6th generation)"
+      case .iPodTouch7: return "iPod touch (7th generation)"
+      case .iPhone4: return "iPhone 4"
+      case .iPhone4s: return "iPhone 4s"
+      case .iPhone5: return "iPhone 5"
+      case .iPhone5c: return "iPhone 5c"
+      case .iPhone5s: return "iPhone 5s"
+      case .iPhone6: return "iPhone 6"
+      case .iPhone6Plus: return "iPhone 6 Plus"
+      case .iPhone6s: return "iPhone 6s"
+      case .iPhone6sPlus: return "iPhone 6s Plus"
+      case .iPhone7: return "iPhone 7"
+      case .iPhone7Plus: return "iPhone 7 Plus"
+      case .iPhoneSE: return "iPhone SE"
+      case .iPhone8: return "iPhone 8"
+      case .iPhone8Plus: return "iPhone 8 Plus"
+      case .iPhoneX: return "iPhone X"
+      case .iPhoneXS: return "iPhone XS"
+      case .iPhoneXSMax: return "iPhone XS Max"
+      case .iPhoneXR: return "iPhone XR"
+      case .iPhone11: return "iPhone 11"
+      case .iPhone11Pro: return "iPhone 11 Pro"
+      case .iPhone11ProMax: return "iPhone 11 Pro Max"
+      case .iPhoneSE2: return "iPhone SE (2nd generation)"
+      case .iPad2: return "iPad 2"
+      case .iPad3: return "iPad (3rd generation)"
+      case .iPad4: return "iPad (4th generation)"
+      case .iPadAir: return "iPad Air"
+      case .iPadAir2: return "iPad Air 2"
+      case .iPad5: return "iPad (5th generation)"
+      case .iPad6: return "iPad (6th generation)"
+      case .iPadAir3: return "iPad Air (3rd generation)"
+      case .iPad7: return "iPad (7th generation)"
+      case .iPad8: return "iPad (8th generation)"
+      case .iPadMini: return "iPad Mini"
+      case .iPadMini2: return "iPad Mini 2"
+      case .iPadMini3: return "iPad Mini 3"
+      case .iPadMini4: return "iPad Mini 4"
+      case .iPadMini5: return "iPad Mini (5th generation)"
+      case .iPadPro9Inch: return "iPad Pro (9.7-inch)"
+      case .iPadPro12Inch: return "iPad Pro (12.9-inch)"
+      case .iPadPro12Inch2: return "iPad Pro (12.9-inch) (2nd generation)"
+      case .iPadPro10Inch: return "iPad Pro (10.5-inch)"
+      case .iPadPro11Inch: return "iPad Pro (11-inch)"
+      case .iPadPro12Inch3: return "iPad Pro (12.9-inch) (3rd generation)"
+      case .iPadPro11Inch2: return "iPad Pro (11-inch) (2nd generation)"
+      case .iPadPro12Inch4: return "iPad Pro (12.9-inch) (4th generation)"
+      case .homePod: return "HomePod"
+      case .simulator(let model): return "Simulator (\(model.safeDescription))"
+      case .unknown(let identifier): return identifier
+      }
+    #elseif os(watchOS)
+      switch self {
+      case .appleWatchSeries0_38mm: return "Apple Watch (1st generation) 38mm"
+      case .appleWatchSeries0_42mm: return "Apple Watch (1st generation) 42mm"
+      case .appleWatchSeries1_38mm: return "Apple Watch Series 1 38mm"
+      case .appleWatchSeries1_42mm: return "Apple Watch Series 1 42mm"
+      case .appleWatchSeries2_38mm: return "Apple Watch Series 2 38mm"
+      case .appleWatchSeries2_42mm: return "Apple Watch Series 2 42mm"
+      case .appleWatchSeries3_38mm: return "Apple Watch Series 3 38mm"
+      case .appleWatchSeries3_42mm: return "Apple Watch Series 3 42mm"
+      case .appleWatchSeries4_40mm: return "Apple Watch Series 4 40mm"
+      case .appleWatchSeries4_44mm: return "Apple Watch Series 4 44mm"
+      case .appleWatchSeries5_40mm: return "Apple Watch Series 5 40mm"
+      case .appleWatchSeries5_44mm: return "Apple Watch Series 5 44mm"
+      case .appleWatchSeries6_40mm: return "Apple Watch Series 6 40mm"
+      case .appleWatchSeries6_44mm: return "Apple Watch Series 6 44mm"
+      case .appleWatchSE_40mm: return "Apple Watch SE 40mm"
+      case .appleWatchSE_44mm: return "Apple Watch SE 44mm"
+      case .simulator(let model): return "Simulator (\(model.safeDescription))"
+      case .unknown(let identifier): return identifier
+      }
+    #elseif os(tvOS)
+      switch self {
+      case .appleTVHD: return "Apple TV HD"
+      case .appleTV4K: return "Apple TV 4K"
+      case .simulator(let model): return "Simulator (\(model.safeDescription))"
+      case .unknown(let identifier): return identifier
+      }
+    #endif
+  }
+
 }
 
-// MARK: - Equatable
+// MARK: Equatable
 extension Device: Equatable {
 
   /// Compares two devices
@@ -1069,7 +1256,7 @@ extension Device: Equatable {
 
 }
 
-// MARK: - Battery
+// MARK: Battery
 #if os(iOS) || os(watchOS)
 @available(iOS 8.0, watchOS 4.0, *)
 extension Device {
@@ -1126,7 +1313,7 @@ extension Device {
 
     /// The user enabled Low Power mode
     public var lowPowerMode: Bool {
-        return ProcessInfo.processInfo.isLowPowerModeEnabled
+      return ProcessInfo.processInfo.isLowPowerModeEnabled
     }
 
     /// Provides a textual representation of the battery state.
@@ -1165,7 +1352,7 @@ extension Device {
 }
 #endif
 
-// MARK: - Device.Batterystate: Comparable
+// MARK: Device.Batterystate: Comparable
 #if os(iOS) || os(watchOS)
 @available(iOS 8.0, watchOS 4.0, *)
 extension Device.BatteryState: Comparable {
@@ -1201,7 +1388,7 @@ extension Device.BatteryState: Comparable {
 
 #if os(iOS)
 extension Device {
-  // MARK: - Orientation
+  // MARK: Orientation
     /**
       This enum describes the state of the orientation.
       - Landscape: The device is in Landscape Orientation
@@ -1223,7 +1410,7 @@ extension Device {
 #endif
 
 #if os(iOS)
-// MARK: - DiskSpace
+// MARK: DiskSpace
 extension Device {
 
   /// Return the root url
@@ -1279,7 +1466,7 @@ extension Device {
 #endif
 
 #if os(iOS)
-// MARK: - Apple Pencil
+// MARK: Apple Pencil
 extension Device {
 
   /**
@@ -1301,7 +1488,7 @@ extension Device {
 
   /// All Apple Pencil Capable Devices
   public static var allApplePencilCapableDevices: [Device] {
-    return [.iPad6, .iPadAir3, .iPad7, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
+    return [.iPad6, .iPadAir3, .iPad7, .iPad8, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4]
   }
 
   /// Returns supported version of the Apple Pencil
@@ -1310,6 +1497,7 @@ extension Device {
       case .iPad6: return .firstGeneration
       case .iPadAir3: return .firstGeneration
       case .iPad7: return .firstGeneration
+      case .iPad8: return .firstGeneration
       case .iPadMini5: return .firstGeneration
       case .iPadPro9Inch: return .firstGeneration
       case .iPadPro12Inch: return .firstGeneration
@@ -1317,6 +1505,9 @@ extension Device {
       case .iPadPro10Inch: return .firstGeneration
       case .iPadPro11Inch: return .secondGeneration
       case .iPadPro12Inch3: return .secondGeneration
+      case .iPadPro11Inch2: return .secondGeneration
+      case .iPadPro12Inch4: return .secondGeneration
+      case .simulator(let model): return model.applePencilSupport
       default: return []
     }
   }
@@ -1324,74 +1515,87 @@ extension Device {
 #endif
 
 #if os(iOS)
-// MARK: - Cameras
+// MARK: Cameras
 extension Device {
 
-  public enum CameraTypes {
+  public enum CameraType {
+    @available(*, deprecated, renamed: "wide")
     case normal
+
+    case wide
     case telephoto
     case ultraWide
   }
 
   /// Returns an array of the types of cameras the device has
-  public var cameras: [CameraTypes] {
+  public var cameras: [CameraType] {
     switch self {
-      case .iPodTouch5: return [.normal]
-      case .iPodTouch6: return [.normal]
-      case .iPodTouch7: return [.normal]
-      case .iPhone4: return [.normal]
-      case .iPhone4s: return [.normal]
-      case .iPhone5: return [.normal]
-      case .iPhone5c: return [.normal]
-      case .iPhone5s: return [.normal]
-      case .iPhone6: return [.normal]
-      case .iPhone6Plus: return [.normal]
-      case .iPhone6s: return [.normal]
-      case .iPhone6sPlus: return [.normal]
-      case .iPhone7: return [.normal]
-      case .iPhoneSE: return [.normal]
-      case .iPhone8: return [.normal]
-      case .iPhoneXR: return [.normal]
-      case .iPad2: return [.normal]
-      case .iPad3: return [.normal]
-      case .iPad4: return [.normal]
-      case .iPadAir: return [.normal]
-      case .iPadAir2: return [.normal]
-      case .iPad5: return [.normal]
-      case .iPad6: return [.normal]
-      case .iPadAir3: return [.normal]
-      case .iPad7: return [.normal]
-      case .iPadMini: return [.normal]
-      case .iPadMini2: return [.normal]
-      case .iPadMini3: return [.normal]
-      case .iPadMini4: return [.normal]
-      case .iPadMini5: return [.normal]
-      case .iPadPro9Inch: return [.normal]
-      case .iPadPro12Inch: return [.normal]
-      case .iPadPro12Inch2: return [.normal]
-      case .iPadPro10Inch: return [.normal]
-      case .iPadPro11Inch: return [.normal]
-      case .iPadPro12Inch3: return [.normal]
-      case .iPhone7Plus: return [.normal, .telephoto]
-      case .iPhone8Plus: return [.normal, .telephoto]
-      case .iPhoneX: return [.normal, .telephoto]
-      case .iPhoneXS: return [.normal, .telephoto]
-      case .iPhoneXSMax: return [.normal, .telephoto]
-      case .iPhone11: return [.normal, .ultraWide]
-      case .iPhone11Pro: return [.normal, .telephoto, .ultraWide]
-      case .iPhone11ProMax: return [.normal, .telephoto, .ultraWide]
+      case .iPodTouch5: return [.wide]
+      case .iPodTouch6: return [.wide]
+      case .iPodTouch7: return [.wide]
+      case .iPhone4: return [.wide]
+      case .iPhone4s: return [.wide]
+      case .iPhone5: return [.wide]
+      case .iPhone5c: return [.wide]
+      case .iPhone5s: return [.wide]
+      case .iPhone6: return [.wide]
+      case .iPhone6Plus: return [.wide]
+      case .iPhone6s: return [.wide]
+      case .iPhone6sPlus: return [.wide]
+      case .iPhone7: return [.wide]
+      case .iPhoneSE: return [.wide]
+      case .iPhone8: return [.wide]
+      case .iPhoneXR: return [.wide]
+      case .iPhoneSE2: return [.wide]
+      case .iPad2: return [.wide]
+      case .iPad3: return [.wide]
+      case .iPad4: return [.wide]
+      case .iPadAir: return [.wide]
+      case .iPadAir2: return [.wide]
+      case .iPad5: return [.wide]
+      case .iPad6: return [.wide]
+      case .iPadAir3: return [.wide]
+      case .iPad7: return [.wide]
+      case .iPad8: return [.wide]
+      case .iPadMini: return [.wide]
+      case .iPadMini2: return [.wide]
+      case .iPadMini3: return [.wide]
+      case .iPadMini4: return [.wide]
+      case .iPadMini5: return [.wide]
+      case .iPadPro9Inch: return [.wide]
+      case .iPadPro12Inch: return [.wide]
+      case .iPadPro12Inch2: return [.wide]
+      case .iPadPro10Inch: return [.wide]
+      case .iPadPro11Inch: return [.wide]
+      case .iPadPro12Inch3: return [.wide]
+      case .iPhone7Plus: return [.wide, .telephoto]
+      case .iPhone8Plus: return [.wide, .telephoto]
+      case .iPhoneX: return [.wide, .telephoto]
+      case .iPhoneXS: return [.wide, .telephoto]
+      case .iPhoneXSMax: return [.wide, .telephoto]
+      case .iPhone11: return [.wide, .ultraWide]
+      case .iPadPro11Inch2: return [.wide, .ultraWide]
+      case .iPadPro12Inch4: return [.wide, .ultraWide]
+      case .iPhone11Pro: return [.wide, .telephoto, .ultraWide]
+      case .iPhone11ProMax: return [.wide, .telephoto, .ultraWide]
       default: return []
     }
   }
 
   /// All devices that feature a camera
   public static var allDevicesWithCamera: [Device] {
-    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
+    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4]
   }
 
   /// All devices that feature a normal camera
+  @available(*, deprecated, renamed: "allDevicesWithWideCamera")
   public static var allDevicesWithNormalCamera: [Device] {
-    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3]
+    return Device.allDevicesWithWideCamera
+  }
+
+  /// All devices that feature a wide camera
+  public static var allDevicesWithWideCamera: [Device] {
+    return [.iPodTouch5, .iPodTouch6, .iPodTouch7, .iPhone4, .iPhone4s, .iPhone5, .iPhone5c, .iPhone5s, .iPhone6, .iPhone6Plus, .iPhone6s, .iPhone6sPlus, .iPhone7, .iPhone7Plus, .iPhoneSE, .iPhone8, .iPhone8Plus, .iPhoneX, .iPhoneXS, .iPhoneXSMax, .iPhoneXR, .iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPhoneSE2, .iPad2, .iPad3, .iPad4, .iPadAir, .iPadAir2, .iPad5, .iPad6, .iPadAir3, .iPad7, .iPad8, .iPadMini, .iPadMini2, .iPadMini3, .iPadMini4, .iPadMini5, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4]
   }
 
   /// All devices that feature a telephoto camera
@@ -1401,7 +1605,7 @@ extension Device {
 
   /// All devices that feature an ultra wide camera
   public static var allDevicesWithUltraWideCamera: [Device] {
-    return [.iPhone11, .iPhone11Pro, .iPhone11ProMax]
+    return [.iPhone11, .iPhone11Pro, .iPhone11ProMax, .iPadPro11Inch2, .iPadPro12Inch4]
   }
 
   /// Returns whether or not the current device has a camera
@@ -1410,8 +1614,14 @@ extension Device {
   }
 
   /// Returns whether or not the current device has a normal camera
+  @available(*, deprecated, renamed: "hasWideCamera")
   public var hasNormalCamera: Bool {
-    return self.cameras.contains(.normal)
+    return self.hasWideCamera
+  }
+
+  /// Returns whether or not the current device has a wide camera
+  public var hasWideCamera: Bool {
+    return self.cameras.contains(.wide)
   }
 
   /// Returns whether or not the current device has a telephoto camera