+++ Tests/Tests.swift
@@ -154,6 +154,7 @@ class DeviceKitTests: XCTestCase {
     XCTAssertTrue(Device.iPhone7.screenRatio == (width: 9, height: 16))
     XCTAssertTrue(Device.iPhone7Plus.screenRatio == (width: 9, height: 16))
     XCTAssertTrue(Device.iPhoneSE.screenRatio == (width: 9, height: 16))
+    XCTAssertTrue(Device.iPhoneSE2.screenRatio == (width: 9, height: 16))
     XCTAssertTrue(Device.iPhone8.screenRatio == (width: 9, height: 16))
     XCTAssertTrue(Device.iPhone8Plus.screenRatio == (width: 9, height: 16))
     XCTAssertTrue(Device.iPhoneX.screenRatio == (width: 9, height: 19.5))
@@ -275,6 +276,17 @@ class DeviceKitTests: XCTestCase {
     XCTAssertEqual(Device.unknown(uuid).description, uuid)
   }
 
+  func testSafeDescription() {
+    for device in Device.allRealDevices {
+      switch device {
+      case .iPhoneXR, .iPhoneXS, .iPhoneXSMax:
+        XCTAssertNotEqual(device.description, device.safeDescription)
+      default:
+        XCTAssertEqual(device.description, device.safeDescription)
+      }
+    }
+  }
+
   func testIsPad() {
     Device.allPads.forEach { XCTAssertTrue($0.isPad) }
   }
@@ -298,6 +310,7 @@ class DeviceKitTests: XCTestCase {
     assertEqualDeviceAndSimulator(device: Device.iPhone7,         property: \Device.ppi, value: 326)
     assertEqualDeviceAndSimulator(device: Device.iPhone7Plus,     property: \Device.ppi, value: 401)
     assertEqualDeviceAndSimulator(device: Device.iPhoneSE,        property: \Device.ppi, value: 326)
+    assertEqualDeviceAndSimulator(device: Device.iPhoneSE2,       property: \Device.ppi, value: 326)
     assertEqualDeviceAndSimulator(device: Device.iPhone8,         property: \Device.ppi, value: 326)
     assertEqualDeviceAndSimulator(device: Device.iPhone8Plus,     property: \Device.ppi, value: 401)
     assertEqualDeviceAndSimulator(device: Device.iPhoneX,         property: \Device.ppi, value: 458)
@@ -340,7 +353,7 @@ class DeviceKitTests: XCTestCase {
   }
 
   func testIsPro() {
-    XCTAssertEqual(Device.allProDevices, [.iPhone11Pro, .iPhone11ProMax, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3])
+    XCTAssertEqual(Device.allProDevices, [.iPhone11Pro, .iPhone11ProMax, .iPadPro9Inch, .iPadPro12Inch, .iPadPro12Inch2, .iPadPro10Inch, .iPadPro11Inch, .iPadPro12Inch3, .iPadPro11Inch2, .iPadPro12Inch4])
   }
 
   func testGuidedAccessSession() {
@@ -387,28 +400,41 @@ class DeviceKitTests: XCTestCase {
 
   func testCameras() {
     for device in Device.allDevicesWithCamera {
-      XCTAssertTrue(device.cameras.contains(.normal) || device.cameras.contains(.telephoto))
+      XCTAssertTrue(device.cameras.contains(.wide) || device.cameras.contains(.telephoto) || device.cameras.contains(.ultraWide))
       XCTAssertTrue(device.hasCamera)
-      XCTAssertTrue(device.hasNormalCamera || device.hasTelephotoCamera)
+      XCTAssertTrue(device.hasWideCamera || device.hasTelephotoCamera || device.hasUltraWideCamera)
     }
     for device in Device.allPhones + Device.allPads + Device.allPods {
       if !Device.allDevicesWithCamera.contains(device) {
-        XCTAssertFalse(device.cameras.contains(.normal))
+        XCTAssertFalse(device.cameras.contains(.wide))
         XCTAssertFalse(device.cameras.contains(.telephoto))
+        XCTAssertFalse(device.cameras.contains(.ultraWide))
         XCTAssertFalse(device.hasCamera)
-        XCTAssertFalse(device.hasNormalCamera)
+        XCTAssertFalse(device.hasWideCamera)
         XCTAssertFalse(device.hasTelephotoCamera)
+        XCTAssertFalse(device.hasUltraWideCamera)
       }
     }
-    for device in Device.allDevicesWithNormalCamera {
-      XCTAssertTrue(device.cameras.contains(.normal))
+    for device in Device.allDevicesWithWideCamera {
+      XCTAssertTrue(device.cameras.contains(.wide))
       XCTAssertTrue(device.hasCamera)
-      XCTAssertTrue(device.hasNormalCamera)
+      XCTAssertTrue(device.hasWideCamera)
     }
     for device in Device.allDevicesWithTelephotoCamera {
       XCTAssertTrue(device.cameras.contains(.telephoto))
       XCTAssertTrue(device.hasCamera)
       XCTAssertTrue(device.hasTelephotoCamera)
+    }
+    for device in Device.allDevicesWithUltraWideCamera {
+      XCTAssertTrue(device.cameras.contains(.ultraWide))
+      XCTAssertTrue(device.hasCamera)
+      XCTAssertTrue(device.hasUltraWideCamera)
+    }
+  }
+
+  func testLidarValues() {
+    for device in Device.allRealDevices {
+      XCTAssertTrue(device.hasLidarSensor == (device == .iPadPro11Inch2 || device == .iPadPro12Inch4))
     }
   }
 
@@ -425,14 +451,20 @@ class DeviceKitTests: XCTestCase {
     XCTAssertEqual(Device.mapToDevice(identifier: "AppleTV6,2").description, "Apple TV 4K")
   }
 
+  func testSafeDescription() {
+    for device in Device.allRealDevices {
+      XCTAssertEqual(device.description, device.safeDescription)
+    }
+  }
+
   /// Test that all the ppi values for applicable devices match the public information available at wikipedia. Test non-applicable devices return nil.
   func testPPI() {
     // Non-applicable devices:
     // Apple TV
-    XCTAssertEqual(Device.appleTV4.ppi, nil)
+    XCTAssertEqual(Device.appleTVHD.ppi, nil)
     XCTAssertEqual(Device.appleTV4K.ppi, nil)
     // Simulators
-    XCTAssertEqual(Device.simulator(Device.appleTV4).ppi, nil)
+    XCTAssertEqual(Device.simulator(Device.appleTVHD).ppi, nil)
     XCTAssertEqual(Device.simulator(Device.appleTV4K).ppi, nil)
   }